implement module format for python code
This script shuffles things around to implement a package structure to ease installation and clarify execution.  Changes include:

 * package structure for code, makes installation a bit easier
 * instructions for install in README
 * reworked bash script which updates geonode, reduced some vexations with hardcoded paths
